Move cli implementation to bootstrap.__main__ module
